requirements.txt
mlflow
psycopg2-binary
torch
transformers
pandas
scikit-learn
dvc
